Server: remove /enter_text_in route
The server should not be responsible for performing multi step actions.

In order for this route to work, it would have to:

1. perform a touch
2. assert the keyboard is showing
3. enter the text

There are many things that can go wrong.  Trying to communicate an error
back to the client will be tedious.  On the client side, there will have
be a lot of logic to consume the errors generated - in effect
duplicating the effort across the client and server.

Resolves: https://github.com/calabash/DeviceAgent.iOS/issues/105