CBXDevice: reports UIDeviceOrientation
This is required for a test on iPhone 3.5 inch form factors.

The layout of views is significantly different on these devices
in landscape orientations so we need to know the orientation to
tell if we need to scroll to a particular view.

I tried to use UIDevice#orientation by registering for orientation
notifications, but I only ever got back orientation unknown.

XCUIDevice seems to return the correct orientation always.